Eliminate duplicate applied LSN check, move directory IO to PartitionStore
